Blackhole: One more rewrite
We generate the blackhole zone as any other zone, except that we
hack nsc.m4 to accept `SOA(@)', so that the same zone file can supply
all blackholed domains.

The consequence is that the blackhole zone is properly versioned
and BIND no longer complains that the SOA is the same, while the
file has changed its timestamp.