When zone contents do not change, version number is not incremented
Replaced direct calls to M4 from the Makefile by calling an auxiliary
script (generated by m4/mkgenzone.m4), which calculates a MD5 hash of
a (suitably normalized) zone file and compares it with a cached hash
of the current version.

The downside is that we run M4 twice on each changed zone. If it ever
causes any harm, we can generate the zone once and move updating of the
version number to the shell script, too.

Based on a patch by Vaclav Ovsik.